Created WordPress types
