+++ src/types/GraphQL.ts
@@ -75,46 +75,3 @@ export type GraphQLResponse = {
     };
   };
 };
-
-type FeaturedMedia = {
-  alt_text: string;
-  caption: string;
-  localFile: {
-    childImageSharp: {
-      fluid: FluidObject;
-    };
-  };
-};
-
-export type WordpressPost = {
-  id: string;
-  title: string;
-  content: string;
-  excerpt: string;
-  date: Date;
-  featured_media: FeaturedMedia;
-};
-
-export type PostTemplate = {
-  title: string;
-  image: {
-    childImageSharp: {
-      fluid: FluidObject;
-    };
-  };
-  image_caption?: string;
-  subtitle?: string;
-  url: string;
-  lang: string;
-  excerpt?: string;
-  description: string;
-  series?: PostSeries;
-  seriesInfo?: PostSeries;
-};
-
-export type SeriesSection = {
-  noDivider?: boolean;
-  title: string;
-  series?: PostSeries;
-  seriesInfo?: PostSeries;
-};
