+++ src/types/index.ts
@@ -2,3 +2,4 @@ export * from './Hash';
 export * from './Algolia';
 export * from './Series';
 export * from './GraphQL';
+export * from './WordPress';
